requirements.txt
# requirements.txt - Local Multimodal RAG (no Pinecone)
torch>=2.4.0
transformers>=4.40.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.10.0
tqdm>=4.66.0
chroma-hnswlib>=0.7.3
chromadb>=0.5.0          # Main Chroma package
open-clip-torch>=2.20.0  # Open-source CLIP
python-dotenv>=1.0.0     # For safe .env loading
google-generativeai>=0.3.0